Extensive code rearrangement.
I've moved a lot of functions (and some struct definitions) around to try to simplify and slim down lime.h. I am also looking forward to introducing a python interface as well as a grid-file format translator, both of which will require some rearrangements.

There are new modules solver.c, grid_aux.c and run.c; I have deleted LTEsolution.c, photon.c, stateq.c and velospline.c. Unused modules have been moved to a subdirectory of that name.

This change will be difficult to check because diff will show huge differences, but the _only_ changes to code are the following:

  - Replaced 'typedef struct {blah} point;' in lime.h with 'struct point {blah};'

  - Replaced 'typedef struct {blah} spec;' in lime.h with 'struct spec {blah};'

  - In the 2 or 3 places where it occurred, replaced 'sizeof(point)' and 'sizeof(spec)' in mallocs with 'sizeof(*<name of the pointer>)'

  - The function run() now returns an integer status. Currently this value is untested, and the only value returned is zero.
