+++ Makefile.defs
@@ -11,35 +11,28 @@ CORESOURCES = \
 	src/collparts.c\
 	src/fastexp.c\
 	src/frees.c\
-	src/getclosest.c\
 	src/grid2fits.c\
 	src/grid.c\
 	src/gridio.c\
-	src/LTEsolution.c\
-	src/magfieldfit.c\
+	src/grid_aux.c\
 	src/messages.c\
 	src/molinit.c\
-	src/photon.c\
 	src/popsin.c\
 	src/popsout.c\
 	src/predefgrid.c\
-	src/ratranInput.c\
 	src/raythrucells.c\
 	src/raytrace.c\
-	src/report.c\
+	src/run.c\
 	src/smooth.c\
+	src/solver.c\
 	src/sourcefunc.c\
-	src/stateq.c\
-	src/statistics.c\
-	src/stokesangles.c\
 	src/tcpsocket.c\
 	src/tree_random.c\
-	src/velospline.c\
-	src/writefits.c\
+	src/writefits.c
 
 STDSOURCES = \
 	src/defaults.c\
-	src/main.c\
+	src/main.c
 
 HDF5SOURCES = \
 	src/grid2hdf5.c
@@ -52,7 +45,7 @@ COREINCLUDES = \
 	src/inpars.h\
 	src/lime.h\
 	src/raythrucells.h\
-	src/tree_random.h\
+	src/tree_random.h
 
 HDF5INCLUDES = \
 	src/grid2hdf5.h
