+++ src/frees.c
@@ -9,6 +9,7 @@
 
 #include "lime.h"
 
+/*....................................................................*/
 void
 freeConfigInfo(configInfo *par){
   int i;
@@ -48,20 +49,7 @@ freeConfigInfo(configInfo *par){
   }
 }
 
-void
-freeInputPars(inputPars *par){
-  free(par->collPartIds);
-  free(par->nMolWeights);
-  free(par->dustWeights);
-  free(par->collPartMolWeights);
-  free(par->moldatfile);
-  free(par->girdatfile);
-  free(par->collPartNames);
-  free(par->gridOutFiles);
-  free(par->gridDensMaxValues);
-  free(par->gridDensMaxLoc);
-}
-
+/*....................................................................*/
 void
 freeGrid(const unsigned int numPoints, const unsigned short numSpecies\
   , struct grid *gp){
@@ -84,19 +72,7 @@ freeGrid(const unsigned int numPoints, const unsigned short numSpecies\
   }
 }
 
-void
-freeGridPointData(const int nSpecies, gridPointData *mol){
-  int i;
-  if(mol!= NULL){
-    for(i=0;i<nSpecies;i++){
-      free(mol[i].jbar);
-      free(mol[i].phot);
-      free(mol[i].vfac);
-      free(mol[i].vfac_loc);
-    }
-  }
-}
-
+/*....................................................................*/
 void
 freeImgInfo(const int nImages, imageInfo *img){
   int i,id;
@@ -112,6 +88,7 @@ freeImgInfo(const int nImages, imageInfo *img){
   free(img);
 }
 
+/*....................................................................*/
 void
 freeMolData(const int nSpecies, molData *md){
   int i,j;
@@ -142,22 +119,7 @@ freeMolData(const int nSpecies, molData *md){
   }
 }
 
-void
-freeMolsWithBlends(struct molWithBlends *mols, const int numMolsWithBlends){
-  int mi, li;
-
-  if(mols != NULL){
-    for(mi=0;mi<numMolsWithBlends;mi++){
-      if(mols[mi].lines != NULL){
-        for(li=0;li<mols[mi].numLinesWithBlends;li++)
-          free(mols[mi].lines[li].blends);
-        free(mols[mi].lines);
-      }
-    }
-    free(mols);
-  }
-}
-
+/*....................................................................*/
 void
 freePopulation(const unsigned short numSpecies, struct populations *pop){
   if(pop != NULL){
@@ -172,6 +134,7 @@ freePopulation(const unsigned short numSpecies, struct populations *pop){
   }
 }
 
+/*....................................................................*/
 void
 freeSomeGridFields(const unsigned int numPoints, const unsigned short numSpecies\
   , struct grid *gp){
