+++ src/lime.h
@@ -180,16 +180,10 @@ typedef struct {
   char molName[80];
 } molData;
 
-/* Data concerning a single grid vertex which is passed from photon() to stateq(). This data needs to be thread-safe. */
-typedef struct {
-  double *jbar,*phot,*vfac,*vfac_loc;
-} gridPointData;
-
-/* Point coordinate */
-typedef struct {
+struct point {
   double x[DIM];
   double xn[DIM];
-} point;
+};
 
 struct rates {
   int t_binlow;
@@ -213,7 +207,7 @@ struct grid {
   double x[DIM], vel[DIM], B[3]; /* B field only makes physical sense in 3 dimensions. */
   double *v1,*v2,*v3;
   int numNeigh;
-  point *dir;
+  struct point *dir;
   struct grid **neigh;
   double *w;
   int sink;
@@ -225,24 +219,18 @@ struct grid {
   struct continuumLine cont;
 };
 
-typedef struct{
-  double x[DIM], xCmpntRay, B[3];
-  struct populations *mol;
-  struct continuumLine cont;
-} gridInterp;
-
-typedef struct {
+struct spec {
   double *intense;
   double *tau;
   double stokes[3];
   int numRays;
-} spec;
+};
 
 /* Image information */
 typedef struct {
   int doline;
   int nchan,trans,molI;
-  spec *pixel;
+  struct spec *pixel;
   double velres;
   double imgres;
   int pxls;
@@ -257,26 +245,6 @@ typedef struct {
   double rotMat[3][3];
   _Bool doInterpolateVels;
 } imageInfo;
-
-struct blend{
-  int molJ, lineJ;
-  double deltaV;
-};
-
-struct lineWithBlends{
-  int lineI, numBlends;
-  struct blend *blends;
-};
-
-struct molWithBlends{
-  int molI, numLinesWithBlends;
-  struct lineWithBlends *lines;
-};
-
-struct blendInfo{
-  int numMolsWithBlends;
-  struct molWithBlends *mols;
-};
 
 /* NOTE that it is assumed that vertx[i] is opposite the face that abuts with neigh[i] for all i.
 */ 
@@ -303,7 +271,7 @@ void	gasIIdust(double,double,double,double *);
 double	gridDensity(configInfo*, double*);
 
 /* More functions */
-void	run(inputPars, image*, const int);
+int	run(inputPars, image*, const int);
 
 _Bool	allBitsSet(const int flags, const int mask);
 _Bool	anyBitSet(const int flags, const int mask);
@@ -311,17 +279,12 @@ _Bool	bitIsSet(const int flags, const int bitI);
 _Bool	onlyBitsSet(const int flags, const int mask);
 
 void	binpopsout(configInfo*, struct grid*, molData*);
-void	calcFastExpRange(const int, const int, int*, int*, int*);
-void	calcGridCollRates(configInfo*, molData*, struct grid*);
-void	calcGridContDustOpacity(configInfo*, const double, double*, double*, const int, struct grid*);
-void	calcGridLinesDustOpacity(configInfo*, molData*, double*, double*, const int, struct grid*);
+void	calcDustData(configInfo*, double*, double*, const double, double*, const int, const double ts[], double*, double*);
+void	calcExpTableEntries(const int, const int);
 void	calcGridMolDensities(configInfo*, struct grid**);
 void	calcGridMolDoppler(configInfo*, molData*, struct grid*);
 void	calcGridMolSpecNumDens(configInfo*, molData*, struct grid*);
-void	calcInterpCoeffs(configInfo*, struct grid*);
-void	calcInterpCoeffs_lin(configInfo*, struct grid*);
 void	calcSourceFn(double, const configInfo*, double*, double*);
-void	calcTableEntries(const int, const int);
 void	checkFirstLineMolDat(FILE *fp, char *moldatfile);
 void	checkGridDensities(configInfo*, struct grid*);
 void	checkUserDensWeights(configInfo*);
@@ -329,65 +292,41 @@ void	copyInparStr(const char*, char**);
 void	delaunay(const int, struct grid*, const unsigned long, const _Bool, const _Bool, struct cell**, unsigned long*);
 void	distCalc(configInfo*, struct grid*);
 double	dotProduct3D(const double*, const double*);
-int	factorial(const int);
 double	FastExp(const float);
 void	fillErfTable();
-void	fit_d1fi(double, double, double*);
-void	fit_fi(double, double, double*);
-void	fit_rr(double, double, double*);
 void	freeConfigInfo(configInfo*);
 void	freeGrid(const unsigned int, const unsigned short, struct grid*);
-void	freeGridPointData(const int, gridPointData*);
 void	freeImgInfo(const int, imageInfo*);
-void	freeInputPars(inputPars*);
 void	freeMolData(const int, molData*);
-void	freeMolsWithBlends(struct molWithBlends*, const int);
 void	freePopulation(const unsigned short, struct populations*);
 void	freeSomeGridFields(const unsigned int, const unsigned short, struct grid*);
 double	gaussline(const double, const double);
 void	getArea(configInfo*, struct grid*, const gsl_rng*);
-void	getclosest(double, double, double, long*, long*, double*, double*, double*);
 double	geterf(const double, const double);
-void	getjbar(int, molData*, struct grid*, const int, configInfo*, struct blendInfo, int, gridPointData*, double*);
 void	getMass(configInfo*, struct grid*, const gsl_rng*);
-void	getVelocities(configInfo *, struct grid *);
-void	getVelocities_pregrid(configInfo *, struct grid *);
-void	gridPopsInit(configInfo*, molData*, struct grid*);
+void	getEdgeVelocities(configInfo *, struct grid *);
 void	input(inputPars*, image*);
+double	interpolateKappa(const double, double*, double*, const int, gsl_spline*, gsl_interp_accel*);
 void	levelPops(molData*, configInfo*, struct grid*, int*, double*, double*, const int);
-void	lineBlend(molData*, configInfo*, struct blendInfo*);
-void	LTE(configInfo*, struct grid*, molData*);
-void	lteOnePoint(molData*, const int, const double, double*);
 void	mallocAndSetDefaultGrid(struct grid**, const size_t, const size_t);
 void	molInit(configInfo*, molData*);
 void	openSocket(char*);
-void	parseInput(inputPars, image*, const int, configInfo*, imageInfo**, molData**);
-void	photon(int, struct grid*, molData*, const gsl_rng*, configInfo*, const int, struct blendInfo, gridPointData*, double*);
 double	planckfunc(const double, const double);
-int	pointEvaluation(configInfo*, const double, double*);
 void	popsin(configInfo*, struct grid**, molData**, int*);
 void	popsout(configInfo*, struct grid*, molData*);
 void	predefinedGrid(configInfo*, struct grid*);
 void	processFitsError(int);
-double	ratranInput(char*, char*, double, double, double);
 void	raytrace(int, configInfo*, struct grid*, molData*, imageInfo*, double*, double*, const int);
 void	readDustFile(char*, double**, double**, int*);
 void	readOrBuildGrid(configInfo*, struct grid**);
-void	readUserInput(inputPars*, imageInfo**, int*, int*);
 unsigned long reorderGrid(const unsigned long, struct grid*);
-void	report(int, configInfo*, struct grid*);
 void	reportInfsAtOrigin(const int, const double*, const char*);
 void	setCollPartsDefaults(struct cpData*);
-void	setUpConfig(configInfo*, imageInfo**, molData**);
 void	setUpDensityAux(configInfo*, int*, const int);
 void	smooth(configInfo*, struct grid*);
 void	sourceFunc_line(const molData*, const double, const struct populations*, const int, double*, double*);
 void	sourceFunc_cont(const struct continuumLine, double*, double*);
 void	sourceFunc_pol(double*, const struct continuumLine, double (*rotMat)[3], double*, double*);
-void	stateq(int, struct grid*, molData*, const int, configInfo*, struct blendInfo, int, gridPointData*, double*, _Bool*);
-void	statistics(int, molData*, struct grid*, int*, double*, double*, int*);
-void	stokesangles(double*, double (*rotMat)[3], double*);
-double	taylor(const int, const float);
 void	writeFitsAllUnits(const int, configInfo*, imageInfo*);
 void	writeGridIfRequired(configInfo*, struct grid*, molData*, const int);
 void	write_VTK_unstructured_Points(configInfo*, struct grid*);
