+++ src/unused/magfieldfit.c
